Added security features for RAPIDAPI key.
